Having only a brief mention of security is problematic. Clarity.
As I mentioned in a reply to [GitHub Issue 109](https://github.com/cozy/cozy-docs/issues/109), the mention of security concerns on this page was a problem. First, I think security should be considered important enough to have its own page or section in the documentation. An invitation to chat on IRC is not good enough.

Other edits are meant improve clarity or make the prose style sound more "natural".